[FE] refactor: 코치 시간 생성 달력과 크루 시간 선택 달력을 분리한다.  (#458)
* refactor: 코치 시간 생성 캘린더 분리

* refactor: 크루 인터뷰 신청 캘린더 분리

* refactor: 코치 시간 생성 캘린더와 크루 인터뷰 신청 캘린더를 children 패턴으로 리팩터링

* refactor: 스타일 로직 분리

* refactor: 필요 없어진 calendar 컴포넌트 삭제

* refactor: 코치 스케줄 캘린더 children 패턴으로 리팩터링

* refactor: 캘린더 관련 필요 없는 파일 제거

* refactor: 캘린더 month와 monthIndex 구분

* refactor: setDay -> removeDate, addOrSetDateBySelectMode 로 분리

* refactor: isSameDate 유틸 함수 활용으로 더 가독성 높이기

* fix: 날짜 범위 버그

* refactor: common 폴더 안에 있는 파일 common 접두사 제거

* refactor: CoachScheduleCalendar -> memo로 감쌈

* refactor: ModalPositionType -> BabyModalPositionType 네이밍 변경

* refactor: isDateNotInOfSelectedDatesAfterToday 함수 네이밍 변경
