+++ frontend/src/components/@common/Calendar/styled.tsx
@@ -1,4 +1,4 @@
-import styled, { css, keyframes } from 'styled-components';
+import styled, { css } from 'styled-components';
 
 export type DayType = 'default' | 'disable' | 'active';
 
@@ -55,106 +55,6 @@ export const MonthPicker = styled.p`
 export const Body = styled.div`
   padding: 1rem;
   font-size: 1.5rem;
-`;
-
-export const WeekDay = styled.div`
-  display: grid;
-  grid-template-columns: repeat(7, 1fr);
-  margin-bottom: 1rem;
-  font-weight: bold;
-
-  div {
-    display: grid;
-    place-items: center;
-    color: ${({ theme }) => theme.colors.gray_150};
-  }
-`;
-
-export const Days = styled.div`
-  display: grid;
-  grid-template-columns: repeat(7, 1fr);
-  gap: 0.2rem;
-`;
-
-const toTop = keyframes`
-  0% {
-    transform: translateY(100%);
-    opacity: 0;
-  }
-  100% {
-    transform: translateY(0);
-    opacity: 1;
-  }
-`;
-
-export const Day = styled.div`
-  position: relative;
-  padding: 0.5rem;
-  animation: ${toTop} 1s;
-  cursor: pointer;
-
-  span {
-    position: absolute;
-  }
-
-  span:nth-of-type(1),
-  span:nth-of-type(3) {
-    width: 2px;
-    height: 0;
-    background-color: ${({ theme }) => theme.colors.pink_50};
-  }
-
-  span:nth-of-type(1) {
-    bottom: 0;
-    left: 0;
-  }
-
-  span:nth-of-type(3) {
-    top: 0;
-    right: 0;
-  }
-
-  span:nth-of-type(2),
-  span:nth-of-type(4) {
-    width: 0;
-    height: 2px;
-    background-color: ${({ theme }) => theme.colors.pink_50};
-  }
-
-  span:nth-of-type(2) {
-    top: 0;
-    left: 0;
-  }
-
-  span:nth-of-type(4) {
-    bottom: 0;
-    right: 0;
-  }
-
-  :hover span {
-    transition: width 0.15s ease-in-out, height 0.15s ease-in-out;
-  }
-
-  :hover span:nth-of-type(1),
-  :hover span:nth-of-type(3) {
-    height: 100%;
-  }
-
-  :hover span:nth-of-type(2),
-  :hover span:nth-of-type(4) {
-    width: 100%;
-  }
-  :hover span:nth-of-type(2) {
-    transition-delay: 0.15s;
-  }
-
-  :hover span:nth-of-type(3) {
-    transition-delay: 0.3s;
-  }
-
-  :hover span:nth-of-type(4) {
-    transition-delay: 0.45s;
-  }
 `;
 
 type MonthContainerProps = {
