+++ frontend/src/components/CoachScheduleCalendar/styled.tsx
@@ -11,6 +11,25 @@ export const Box = styled.div`
   border-radius: 20px;
 `;
 
+export const WeekDay = styled.div`
+  display: grid;
+  grid-template-columns: repeat(7, 1fr);
+  margin-bottom: 1rem;
+  font-weight: bold;
+
+  div {
+    display: grid;
+    place-items: center;
+    color: ${({ theme }) => theme.colors.gray_150};
+  }
+`;
+
+export const Days = styled.div`
+  display: grid;
+  grid-template-columns: repeat(7, 1fr);
+  gap: 0.2rem;
+`;
+
 type CalendarDayProps = {
   type?: DayType;
   today?: boolean;
