+++ frontend/src/context/CalendarProvider.tsx
@@ -2,6 +2,8 @@ import { createContext, useContext, useState } from 'react';
 
 import { SelectMode } from '@/types/domain';
 
+import { convertMonthIndexToMonth, convertMonthToMonthIndex } from '@/utils';
+
 type CalendarProviderProps = {
   selectMode: SelectMode;
   children: React.ReactNode;
@@ -26,7 +28,8 @@ type CalendarActionsType = {
   handleClickPrevMonth: () => void;
   handleClickNextMonth: () => void;
   handleClickMonthPicker: () => void;
-  setDay: (day: number) => void;
+  removeDate: (day: number) => void;
+  addOrSetDateBySelectMode: (day: number) => void;
   getHandleClickMonth: (monthIndex: number) => () => void;
   resetSelectedDates: () => void;
   addSelectedDates: (dates: Date[]) => void;
@@ -76,28 +79,32 @@ const CalendarProvider = ({ selectMode, children }: CalendarProviderProps) => {
   const date = new Date();
 
   const [year, setYear] = useState(date.getFullYear());
-  const [month, setMonth] = useState(date.getMonth());
+  const [month, setMonth] = useState(date.getMonth() + 1);
 
   const [selectedDates, setSelectedDates] = useState<Date[]>([]);
 
   const [showMonthPicker, setShowMonthPicker] = useState(false);
 
-  const firstDay = new Date(year, month, 1).getDay();
+  const firstDay = new Date(year, convertMonthToMonthIndex(month), 1).getDay();
   const daysOfMonth = [31, getFebruaryDays(year), 31, 30, 31, 30, 31, 31, 30, 31, 30, 31];
-  const daysLength = daysOfMonth[month] + firstDay;
+  const daysLength = daysOfMonth[convertMonthToMonthIndex(month)] + firstDay;
 
   const isSameDate = (date: Date, day: number) =>
-    date.year === year && date.month === month + 1 && date.day === day;
+    date.year === year && date.month === month && date.day === day;
 
   const isToday = (day: number) => {
     const date = new Date();
 
-    return day === date.getDate() && year === date.getFullYear() && month === date.getMonth();
+    return (
+      day === date.getDate() &&
+      year === date.getFullYear() &&
+      convertMonthToMonthIndex(month) === date.getMonth()
+    );
   };
 
   const isBelowToday = (day: number) => {
     const today = new Date().getTime();
-    const date = new Date(year, month, day).getTime();
+    const date = new Date(year, convertMonthToMonthIndex(month), day).getTime();
 
     return date <= today;
   };
@@ -126,10 +133,10 @@ const CalendarProvider = ({ selectMode, children }: CalendarProviderProps) => {
     },
     handleClickPrevMonth() {
       setMonth((prev) => {
-        if (prev === 0) {
+        if (prev === 1) {
           actions.handleClickPrevYear();
 
-          return 11;
+          return 12;
         }
 
         return prev - 1;
@@ -137,31 +144,28 @@ const CalendarProvider = ({ selectMode, children }: CalendarProviderProps) => {
     },
     handleClickNextMonth() {
       setMonth((prev) => {
-        if (prev === 11) {
+        if (prev === 12) {
           actions.handleClickNextYear();
 
-          return 0;
+          return 1;
         }
 
         return prev + 1;
       });
     },
-    handleClickMonthPicker: () => {
+    handleClickMonthPicker() {
       setShowMonthPicker(true);
     },
-    setDay: (day: number) => {
-      if (isSelectedDate(day)) {
-        setSelectedDates((prev) => prev.filter((date) => !isSameDate(date, day)));
-
-        return;
-      }
-
+    removeDate(day: number) {
+      setSelectedDates((prev) => prev.filter((date) => !isSameDate(date, day)));
+    },
+    addOrSetDateBySelectMode(day: number) {
       selectMode === 'SINGLE'
-        ? setSelectedDates([{ year, month: month + 1, day }])
-        : setSelectedDates((prev) => [...prev, { year, month: month + 1, day }]);
+        ? setSelectedDates([{ year, month, day }])
+        : setSelectedDates((prev) => [...prev, { year, month, day }]);
     },
     getHandleClickMonth: (monthIndex: number) => () => {
-      setMonth(monthIndex);
+      setMonth(convertMonthIndexToMonth(monthIndex));
       setShowMonthPicker(false);
     },
     resetSelectedDates() {
